Validation exception stacktrace shortened
