+++ src/jvm/test/org/jetbrains/kotlinx/lincheck/test/representation/ValidateFunctionTest.kt
@@ -1,0 +1,48 @@
+/*
+ * Lincheck
+ *
+ * Copyright (C) 2019 - 2023 JetBrains s.r.o.
+ *
+ * This Source Code Form is subject to the terms of the
+ * Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
+ * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.jetbrains.kotlinx.lincheck.test.representation
+
+import org.jetbrains.kotlinx.lincheck.annotations.*
+import org.jetbrains.kotlinx.lincheck.annotations.Operation
+import org.jetbrains.kotlinx.lincheck.test.util.*
+import org.jetbrains.kotlinx.lincheck.verifier.*
+import org.junit.*
+import java.util.concurrent.atomic.*
+
+class ValidateFunctionTest : VerifierState() {
+    val c = AtomicInteger()
+
+    @Operation
+    fun inc() = c.incrementAndGet()
+
+    override fun extractState() = c.get()
+
+    @Validate
+    fun validateNoError() {
+    }
+
+    var validateInvoked: Int = 0
+
+    // This function fails on the 5ht invocation
+    @Validate
+    fun validateWithError() {
+        validateInvoked++
+        if (validateInvoked == 5) error("Validation works!")
+    }
+
+    @Test
+    fun test() = runModelCheckingTestAndCheckOutput("validation_function_failure.txt") {
+        iterations(1)
+        invocationsPerIteration(1)
+        actorsBefore(3)
+        actorsAfter(10)
+    }
+
+}
